devproc: fix proccrlmemio bugs
dont kill the calling process when demand load fails if fixfault()
is called from devproc. this happens when you delete the binary
of a running process and try to debug the process accessing uncached
pages thru /proc/$pid/mem file.

fixes to procctlmemio():

- fix missed unlock as txt2data() can error
- make sure the segment isnt freed by taking a reference (under p->seglock)
- access the page with segment locked (see comment)
- get rid of the segment stealer lock

other stuff:

- move txt2data() and data2txt() to segment.c
- add procpagecount() function
- make return type mcounseg() to ulong

+++ sys/src/9/port/proc.c
@@ -1510,7 +1510,7 @@ killbig(char *why)
 			s = p->seg[i];
 			if(s == nil || !canqlock(s))
 				continue;
-			l += (ulong)mcountseg(s);
+			l += mcountseg(s);
 			qunlock(s);
 		}
 		qunlock(&p->seglock);
